# Conflicts:
#	backend/ziip/src/main/java/com/lastdance/ziip/global/exception/validator/MemberValidator.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/controller/MemberController.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/LoginDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/TokenDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/request/MemberInfoUpdateRequestDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/request/NicknameRequestDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/response/LoginResponseDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/response/MemberInfoResponseDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/response/MemberUpdateResponseDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/response/NickNameResponseDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/dto/response/RefreshTokenResponseDto.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/repository/MemberRepository.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/service/MemberService.java
#	backend/ziip/src/main/java/com/lastdance/ziip/member/service/MemberServiceImpl.java
#	backend/ziip/src/main/resources/application.yml

+++ backend/ziip/src/main/java/com/lastdance/ziip/diary/repository/entity/DiaryComment.java
@@ -2,12 +2,10 @@ package com.lastdance.ziip.diary.repository.entity;
 
 import com.lastdance.ziip.global.entity.BaseEntity;
 import com.lastdance.ziip.member.repository.entity.Member;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
+
+import javax.persistence.*;
+import javax.validation.constraints.NotNull;
+
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -22,16 +20,20 @@ import lombok.NoArgsConstructor;
 public class DiaryComment extends BaseEntity {
 
     @Id @GeneratedValue
-    private Integer id;
+     
+    private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
+     
     private Member member;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "diary_id")
+     
     private Diary diary;
 
+     
     private String content;
 
 }
