+++ backend/ziip/src/main/java/com/lastdance/ziip/diary/repository/entity/DiaryPhoto.java
@@ -1,12 +1,10 @@
 package com.lastdance.ziip.diary.repository.entity;
 
 import com.lastdance.ziip.global.entity.BaseEntity;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
+
+import javax.persistence.*;
+import javax.validation.constraints.NotNull;
+
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -21,13 +19,18 @@ import lombok.NoArgsConstructor;
 public class DiaryPhoto extends BaseEntity {
 
     @Id @GeneratedValue
-    private Integer id;
+     
+    private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "diary_id")
+     
     private Diary diary;
 
+     
     private String imgUrl;
+
+     
     private String imgName;
 
 }
